Move Poms::Media and Poms::URIs to Poms::API submodule
