--- lib/poms/uris.rb
@@ -1,28 +1,0 @@
-require 'addressable/uri'
-
-# This module builds URI objects for various endpoints of the POMS API
-module Poms
-  module URIs
-    BASE_PARAMS = { scheme: 'https', host: 'rs.poms.omroep.nl' }
-
-    module Media
-      def self.single(mid)
-        uri_for_path("/#{mid}")
-      end
-
-      def self.multiple
-        uri_for_path("/multiple")
-      end
-
-      def self.descendants(mid)
-        uri_for_path("/#{mid}/descendants")
-      end
-
-      private
-
-      def self.uri_for_path(path)
-        Addressable::URI.new(BASE_PARAMS.merge(path: "/v1/api/media#{path}"))
-      end
-    end
-  end
-end
